chore: add cargo_check_verbose mod
